Re-factor filtering of data
