+++ tabs/aggregate_summary.py
@@ -1,0 +1,27 @@
+import streamlit as st
+import pandas as pd
+import numpy as np
+
+from utils import rounded_dollars_md
+from streamlit_utils import graph_selected_col
+
+
+def tab_aggregate_content(
+        df,
+        selected_date,
+):
+    # Perform all calcs on the data in the filtering function first, then sue the data to present it here.
+
+    # # Get data for current month
+    # df_current = df[]
+    # selected_date
+    # st.write("# Market Players")
+    # # Group by 'Date' and count unique names
+    # market_player_count_df = df.groupby(date_column)[category_column].nunique()
+    # st.write(market_player_count_df)
+# 
+# 
+    # st.write(df)
+
+
+    return 0
