--- tabs/tab1.py
@@ -1,23 +1,0 @@
-import streamlit as st
-
-def tab1_content():
-    st.write("this is tab 1")
-
-    st.radio('Select one:', [1, 2])
-    st.button('Hit me')
-    # st.data_editor('Edit data', data_dict['bb_df'])
-    st.checkbox('Check me out')
-    st.radio('Pick one:', ['nose','ear'])
-    st.selectbox('Select', [1,2,3])
-    st.multiselect('Multiselect', [1,2,3])
-    st.slider('Slide me', min_value=0, max_value=10)
-    st.select_slider('Slide to select', options=[1,'2'])
-    st.text_input('Enter some text')
-    st.number_input('Enter a number')
-    st.text_area('Area for textual entry')
-    st.date_input('Date input')
-    st.time_input('Time entry')
-    st.file_uploader('File uploader')
-    # st.download_button('On the dl', data_dict['bb_df'])
-    st.camera_input("一二三,茄子!")
-    st.color_picker('Pick a color')
